requirements.txt
# requirements.txt - PRODUCTION READY FOR RENDER

# Core ML & Data Science
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.7.2
scipy==1.15.3
xgboost
imbalanced-learn

# Model Interpretability
shap

# Visualization (optional for production, but keep if needed)
matplotlib==3.10.6
seaborn==0.13.2

# API Framework
Flask==3.1.2
Werkzeug==3.1.1

# 🔥 NEW: Production WSGI server (CRITICAL for Render)
gunicorn==21.2.0

# Configuration
python-dotenv==1.1.1

# MLflow (optional - only if using MLflow model registry)
mlflow==3.4.0

# Utilities
joblib==1.5.2
packaging==25.0
python-dateutil==2.9.0.post0
pytz==2025.2

# Type Checking
typing_extensions==4.15.0